Added NuGet package manager and JSON .NET package.
Added NuGet package manager to the project which allows for package management
as well as the first package, JSON .NET so that JSON can be serialised and
deserialised on command.

A small test script was added to a test cube that should print three debug
messages should JSON be functioning correctly.