Added skeleton classes for all of the networking packets.
All network packets as designated in Issues #3 and #4 have been implemented as
empty classes. They all derive from an abstract parent class Packet which
handles serialisation and deserialisation of packets. Packet serialisation works
great and handles itself out of the box. Deserialisation on the other hand is
going to need some work to parse and cast the object into the concrete
class-types.