Added a packet socket class, more socket functions.
Added a wrapper for the native websocket class that will handle all packet
interaction. Currently undocumented.

A few more functions were additionally added, this time in the form of DoX which
are intented as endpoints for the client to send data to the server, rather than
receiving data which is done by the OnX set of event callbacks.