Updated strings and prefabs, added netcode for manual IP lobby connect.
Strings were fixed because they had spaces in their names.
Prefabs were updated to accomodate a few changes I made to the UI.

PacketSocket.cs was update for the manual IP connector in the main menu which
will now attempt to successfully stage a connection before opening the lobby
list, and if it fails, it will show an error message.