Fixed camera move on pause. Added infrastructure for online server connections.
The main server can now be connected to by first posting a HTTP POST connection
via. a login interface to get the client ID and token. Logic was added so that
the local servers and online servers each provide different data to the lobby
socket (this not yet tested or working with the game socket for the moment).

New prefab too for login interface. Takes a username and password.

The camera no longer pans around the scene while any menu is open now too which
fixes a problem that has been occuring for a while with the pause menu.